Pull form out of TripForm, move back to CreateNewTrip only
